+++ iocore/dns/P_DNSProcessor.h
@@ -68,7 +68,7 @@ extern unsigned int dns_sequence_number;
 #define DNS_SEQUENCE_NUMBER_RESTART_OFFSET  4000
 #define DNS_PRIMARY_RETRY_PERIOD            HRTIME_SECONDS(5)
 #define DNS_PRIMARY_REOPEN_PERIOD           HRTIME_SECONDS(60)
-#define BAD_DNS_RESULT                      ((HostEnt*)(long)-1)
+#define BAD_DNS_RESULT                      ((HostEnt*)(uintptr_t)-1)
 #define DEFAULT_NUM_TRY_SERVER              8
 
 // these are from nameser.h
@@ -306,7 +306,7 @@ TS_INLINE DNSHandler::DNSHandler()
   n_con(0),
   options(0),
   in_flight(0), name_server(0), in_write_dns(0), hostent_cache(0), last_primary_retry(0), last_primary_reopen(0),
-  m_res(0), txn_lookup_timeout(0), generator(time(NULL) ^ (long) this)
+  m_res(0), txn_lookup_timeout(0), generator((uint32)((uintptr_t)time(NULL) ^ (uintptr_t)this))
 {
   for (int i = 0; i < MAX_NAMED; i++) {
     ifd[i] = -1;
