+++ iocore/dns/P_SplitDNSProcessor.h
@@ -167,7 +167,7 @@ TS_INLINE bool SplitDNSConfig::isSplitDNSEnabled()
    -------------------------------------------------------------- */
 struct DNSServer
 {
-  unsigned long x_server_ip[MAXNS];
+  unsigned int x_server_ip[MAXNS];
   char x_dns_ip_line[MAXDNAME * 2];
 
   char x_def_domain[MAXDNAME];
